Add files via upload
BTC Reward contract -- MainContract.sol is setup for 4 dev wallets split evenly, MainContract2Wallet.sol is setup for 2 dev wallets split evenly.

The contract is mostly commented, and things can be changed as needed but, I won't be going into too much detail here on how to change things or how to use the contract. The main things to change are labeled as such.